Fixes to VC++ compiler errors.
git-svn-id: https://crag.googlecode.com/svn/trunk@971 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/gfx/Renderer.cpp
@@ -345,7 +345,7 @@ Color4f Renderer::CalculateLighting(Vector3 const & position) const
 		Vector3 light_position = light.GetModelViewTransformation().GetTranslation();
 		
 		Vector3 frag_to_light = light_position - position;
-		float distance_squared = LengthSq(frag_to_light);
+		float distance_squared = static_cast<float>(LengthSq(frag_to_light));
 		
 		float attenuation = Clamped(1.f / distance_squared, 0.f, 1.f);
